pyproject.toml
[project]
name = "proyectoiabd"
version = "0.1.0"
description = "Proyecto Curso Especialización IABD"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "folium>=0.20.0",
    "numpy>=2.4.4",
    "ortools>=9.15.6755",
    "pandas>=3.0.2",
    "pip-system-certs>=5.3",
    "pyarrow>=24.0.0",
    "pydantic>=2.13.3",
    "pyspark==3.5.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "setuptools>=82.0.1",
    "streamlit>=1.56.0",
    "streamlit-folium>=0.27.1",
    "uvicorn>=0.46.0",
]


[tool.uv]
package = true
